fix: dependency issue: constants and types imported into the user entity made it so I couldn't import user entity into the status service
