+++ src/backend/app/src/status/status.types.ts
@@ -1,0 +1,19 @@
+// NULL or MISSING === OFFLINE
+export type UserState =
+  | 'OFFLINE'
+  | 'ONLINE'
+  | 'SEARCHING'
+  | 'PLAYING'
+  | 'OBSERVING'
+  | 'CHALLENGED'
+  | 'CONNECTION_DENIED';
+
+export const USER_STATES: UserState[] = [
+  'OFFLINE',
+  'ONLINE',
+  'SEARCHING',
+  'PLAYING',
+  'OBSERVING',
+  'CHALLENGED',
+  'CONNECTION_DENIED',
+];
